Refactor code
- Move classes into it's own file
- Fix nil errors
- Add TODO comments
