+++ lib/solargraph/parser/parser_gem/node_processors/casgn_node.rb
@@ -29,6 +29,8 @@ module Solargraph
             process_children
           end
 
+          # TODO: Move this out of [CasgnNode] once [Solargraph::Parser::NodeProcessor] supports
+          # multiple processors.
           def process_struct_assignment
             processor_klass = Convention::StructDefinition::NodeProcessors::StructNode
             processor = processor_klass.new(node, region, pins, locals)
